requirements.txt
streamlit
pandas
matplotlib
requests
face_recognition
numpy
Pillow
deepface